Bug #2231: Hostgroup deletion follows :restrict orphan policy
Hostgroup deletion is restricted to hostgroups without children

Hostgroup tests are rewritten to follow the :restrict orphan policy

Wrap AncestryException in ForemanException

Nested deletions throw flash error on UI

Add restrict control message on the API

Fix test to work with new orphan policy

Fix comment in test

Check exceptions for :restrict deletion of nested hostgroups in API v2

+++ app/controllers/api/v1/hostgroups_controller.rb
@@ -64,7 +64,11 @@ module Api
       param :id, :identifier, :required => true
 
       def destroy
-        process_response @hostgroup.destroy
+        if @hostgroup.has_children?
+          render :json => {'message'=> _("Cannot delete group %{current} because it has nested groups.") % { :current => @hostgroup.label } }, :status => :conflict
+        else
+          process_response @hostgroup.destroy
+        end
       end
 
     end
